+++ packages/flusmic/lib/src/models/document/document.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: invalid_annotation_target
+
 import 'package:flusmic/src/models/info/alternate_language.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
@@ -17,9 +19,12 @@ class Document<T> with _$Document<T> {
     required String type,
     @JsonKey(name: 'alternate_languages')
         List<AlternateLanguage>? alternateLanguages,
-    @JsonKey(name: 'first_publication_date') String? firstPublicationDate,
-    @JsonKey(name: 'last_publication_date') String? lastPublicationDate,
-    @JsonKey(name: 'linked_documents') List<String>? linkedDocuments,
+    @JsonKey(name: 'first_publication_date')
+        String? firstPublicationDate,
+    @JsonKey(name: 'last_publication_date')
+        String? lastPublicationDate,
+    @JsonKey(name: 'linked_documents')
+        List<String>? linkedDocuments,
     T? data,
     List<String>? slugs,
     String? href,
@@ -32,4 +37,11 @@ class Document<T> with _$Document<T> {
     T Function(Object? json) fromJsonT,
   ) =>
       _$DocumentFromJson(json, fromJsonT);
+
+  // coverage:ignore-start
+  ///Temporary fix to include Document in custom model
+  // ignore: annotate_overrides
+  Map<String, dynamic> toJson(Object? Function(T) toJsonT) =>
+      super.toJson(toJsonT);
+  // coverage:ignore-end
 }