+++ packages/flusmic/lib/src/models/response/flusmic_response.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: invalid_annotation_target
+
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'flusmic_response.freezed.dart';